+++ docs/version_history.md
@@ -17,6 +17,7 @@ date
 * fix: bug in prefight quotes not being said by second participant
 * feat: fury, another (hopefully) fun game mechanic
 * feat: fury-related techniques and styles
+* feat: new fight AI that rushes when fights against 4-distance moves or when has a bigger crowd
 
 
 ### v.0.6.4-beta